services/data_ingestion_service/requirements.txt
chromadb==1.0.10
fastapi==0.115.9
httpx==0.28.1
langchain==0.3.25
langchain-community==0.3.24
llama-index==0.12.37
llama-index-embeddings-langchain==0.3.0
pydantic==2.11.4
pydrive2==1.21.3
python-multipart==0.0.20
python-pptx==1.0.2
sentence-transformers==4.1.0
sqlalchemy==2.0.41
uvicorn==0.34.2
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.7.0+cpu
langchain-docling==0.2.0
docling==2.33.0
pymupdf4llm==0.0.24
pymupdf==1.26.0
